// public/apps/the_peoples_edh_bracket/the-peoples-edh-bracket-client/src/types/bindings/TrackedDeckWithAnalysis.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Deck } from "./Deck";
import type { PointsHistogramBucket } from "./PointsHistogramBucket";

export type TrackedDeckWithAnalysis = { uuid: string, tracker_person_uuid: string, name: string, url_source: string | null, created_at: string, updated_at: string | null, deck: Deck, total_points: string, histogram: Array<PointsHistogramBucket>, };